Build against more recent Rubies
Ruby 2.5 became unsupported (end of life) on 2021-03-31:

  https://www.ruby-lang.org/en/downloads/branches/

Ruby 3 requires more of an update to handle the new way that keyword
arguments are managed. Let's deal with that in a later version.

Specify latest Rails in gemfiles (for build): this isn't strictly
necessary at the moment, since when Travis builds the app it's always
fetching the latest version of the gem, but it feels nice to specify
what the latest expected versions are.

The sudo option is deprecated and doesn't do anything anymore