Remove bundler-audit from codeclimate tests
We have much better ways of identifying vulnerabilities these days
through github, and in any case this fails because codeclimate requires
a Gemfile.lock and ours isn't committed (yet? I'm not sure if it should
be or not).

Also regenerate and encrypt the codeclimate token. Apparently it's not
essential that this should be private, but it's easy to do, so why not.