Build/test against Ruby 3.0
We've been passing translate options in a way which is not compatible
with Ruby 3. This led to the following error:

    ActionView::Template::Error: wrong number of arguments (given 2, expected 1)

Ruby 3 interprets keyword arguments vs. hashes differently. The I18n
source code has a comment which shows the correct expected usage:

  https://github.com/ruby-i18n/i18n/blob/v1.8.10/lib/i18n.rb#L181-L194

Removing the {} causes the arguments to be sent correctly.

Note that Ruby 3 is not expected to work with Rails 5.2:

  https://github.com/rails/rails/issues/40938#issuecomment-751357907