Prevent intermittent test failures with teardown
The issue is that this test creates a `fakemin` directory which itself
includes tests. Depending on test order, some of the time, that fakemin
directory is left after the test run, so minitest tries to include those
tests in the test run, but they're not in the right context.

This is the error we get:

    NameError: uninitialized constant FakeminTest::Fakemin
        /Users/dgmstuart/dev/godmin/test/tmp/fakemin/test/fakemin_test.rb:5:in `block in <class:FakeminTest>'