Only look for pundit policies in the engine
I believe this fixes a regression between Godmin 1.x and Godmin 2.x
regarding the engine installation: the previous behaviour was that
Godmin only looked for Policy classes in the main app.

In Godmin 2.0 controller actions correctly find the policy class in the
engine because the whole controller class is namespaced under the
engine, but direct calls to `policy` in views are not namespaced, so
result in a policy class from the main application being referenced,
even if one exists in the engine.

Background
==========
The Godmin v1 we had the following logic:

    module Authorization
      ...

      def policy(record)
        policies[record] ||= PolicyFinder.find(record, engine_wrapper.namespace).new(admin_user, record)
      end

      ...
    end

    module Authorization
      class PolicyFinder
        class << self
          def find(object, namespace = nil)
            klass = ...

            if namespace
              "#{namespace}::#{klass}Policy"
            else
              "#{klass}Policy"
            end.constantize
          end
        end
      end
    end

...and model classes were always referenced from the main application.

The effect of which was that when run as an engine, Godmin _always_
looked for policy classes within the engine namespace.

Example: if the model was `Widget`, and the engine was called `Admin`,
then Godmin would look for `Admin::WidgetPolicy`, overriding any policy
which might exist in the main application

Regression
==========
When Godmin v2 was created and models were moved into the engine (eg.
Admin::Widget) this was mostly no longer necessary: Pundit mostly acts
on controller actions, and so are already nested under the engine's
namespace (eg. `Admin::WidgetsController`).

However, there are a number of ways in which direct calls to `policy`
could result in Godmin looking for policies in the main application.

Examples:

- `policy(::Widget)` within the `Admin` namespace
- `policy(Widget)` within a view (ie. not namespaced)
- `policy(Admin::Factory.widgets.first)` anywhere

Particularly with the last one, I don't think there's a way around this:
forcing the resulting `Widget` to be an instance of `Admin::Widget`
instead of `::Widget` feels like it would require a level of
monkeypatching and/or metaprogramming which I don't feel like we should
be getting into.

Solution
========
The approach involves using the namespaced policies feature of Pundit:

  https://github.com/varvet/pundit/tree/v2.1.0#policy-namespacing
  https://github.com/varvet/pundit/blob/v2.1.0/lib/pundit/policy_finder.rb#L75-L78

We add the engine namespace on the front of the record before sending to
Pundit to find the policy. This has the effect of always looking for
policies in the engine namespace.

We then need to do some work to handle a number of cases:
- Where the record is already namespaced, because we're referring to the
  version of the model within the engine (eg. Admin::Widget)
- Where the record is already namespaced because it's part of another
  engine which is part of the admin site
- Where the record has the same initial string as the engine, but is not
  namespaced (eg. AdminNotification)

To be honest, this string matching on class namespaces feels like a bit
of a hack, but there doesn't seem to be a more idiomatic way to achieve
the result we're looking for.

Note: not all the cases are covered in tests - eg. odd edge cases where
there's an ActiveModel instance which has a `model_name` which is
different from the class name (?)

Impact
======
Now when using the Pundit mixins, or the `policy` method, Godmin will
only look for policies under the engine namespace.

Calls to `policy` will fail if there's a policy which only exists in
the main app, but I believe that's what we want: the admin engine is a
separate application which just happens to use the same model layer as
the main application.

Also I believe this was more or less the behaviour of Godmin 1, though
it's hard to be sure.

In complex admin applications where the admin application itself is
split up into multiple engines (eg. admin, admin-mailing, admin-users),
these now need to be namespaced under the same namespace as the engine
(ie. they must be eg. Admin::Content, rather than Content::Admin or some
other variation), or else all the policies need to live within the main
Godmin engine. This is because Godmin will now try to prepend the engine
namespace for all policy lookups unless the namespace is there already.

I believe that this change removes the necessity to have a class inside
the engine for every model, mirroring the models in the main app, but
that can be removed as a future change.