Remove README comment about stable versions
To be honest it's not really clear what the stability status of Godmin
is. 2.0 was done somewhat hastily in order to get compatibility with
Rails 6, and as a result some experimental changes were introduced.

It hasn't been possible to get the 1.5 branch building successfully on
recent ruby versions, so I don't think it's something we can commit to
maintaining.