Fix: column_value returns values
After releasing v2.2.0 and running godmin on Rails 5, column values no
longer rendered unless there was a partial or a function which governs
the behaviour of that column.

The issue seems to be that the partial_override block was always
returning nil, when in the default case (no partial is defined) it was
expected to return the column_value.

This is one of those "I don't know how this ever worked" problems: none
of the files involved seem to have been touched in years, and I'd be
surprised if this was a result of one of the Ruby or Rails upgrades
which were made on the project I experienced the issue on.

In any case, this seems to fix the issue, and columns with defined
partials also seem to work as expected.