Bump version to 2.3.0
Use the latest Pundit version.

TL;DR: load order is complicated

Godmin was not usable with the previous Pundit version because that
introduced a deprecation warning which relied on ActiveSupport. Due to
the way that gem loading works, this caused problems with Godmin, since
the relevant part of ActiveSupport was not loaded when Pundit tried to
show the deprecation message.

This should in the worst case result in deprecation warnings being shown
for projects which want to upgrade Godmin, but without making the
(trivial) change from `include Pundit` to `include
Pundit::Authorization`.